+++ packages/adp-tooling/src/writer/writer-config.ts
@@ -5,9 +5,9 @@ import type { AbapServiceProvider } from '@sap-ux/axios-extension';
 import { FlexLayer } from '../types';
 import { getProviderConfig } from '../abap';
 import { getCustomConfig } from './project-utils';
-import type { AdpWriterConfig, ConfigAnswers } from '../types';
+import type { AdpWriterConfig, AttributesAnswers, ConfigAnswers, UI5Version } from '../types';
 import { getNewModelEnhanceWithChange } from './descriptor-content';
-import { UI5VersionInfo, getFormattedVersion, getOfficialBaseUI5VersionUrl } from '../ui5';
+import { getFormattedVersion, getLatestVersion, getOfficialBaseUI5VersionUrl, getVersionToBeUsed } from '../ui5';
 
 interface ConfigOptions {
     /**
@@ -19,22 +19,21 @@ interface ConfigOptions {
      */
     configAnswers: ConfigAnswers;
     /**
+     * User-provided project attribute answers.
+     */
+    attributeAnswers: AttributesAnswers;
+    /**
      * The FlexLayer indicating the deployment layer (e.g., CUSTOMER_BASE or VENDOR).
      */
     layer: FlexLayer;
     /**
-     * Default project parameters.
-     */
-    defaults: {
-        /**
-         * The default namespace for the project.
-         */
-        namespace: string;
-    };
-    /**
      * The package.json information used to generate custom configuration.
      */
     packageJson: Package;
+    /**
+     * Public UI5 Versions.
+     */
+    publicVersions: UI5Version;
     /**
      * Logger instance for debugging and error reporting.
      */
@@ -55,7 +54,7 @@ interface ConfigOptions {
  * @returns {Promise<AdpWriterConfig>} A promise that resolves to the generated ADP writer configuration.
  */
 export async function getConfig(options: ConfigOptions): Promise<AdpWriterConfig> {
-    const { configAnswers, defaults, layer, logger, packageJson, provider } = options;
+    const { configAnswers, attributeAnswers, layer, logger, packageJson, provider, publicVersions } = options;
     const ato = await provider.getAtoInfo();
     const operationsType = ato.operationsType ?? 'P';
 
@@ -65,15 +64,22 @@ export async function getConfig(options: ConfigOptions): Promise<AdpWriterConfig
     const isCloudProject = await provider.isAbapCloud();
     const isCustomerBase = layer === FlexLayer.CUSTOMER_BASE;
 
-    const ui5Info = UI5VersionInfo.getInstance(layer);
-    const ui5Version = isCloudProject ? ui5Info.getLatestVersion() : ui5Info.getVersionToBeUsed('', isCustomerBase);
+    const ui5Version = isCloudProject
+        ? getLatestVersion(publicVersions)
+        : getVersionToBeUsed(attributeAnswers.ui5Version, isCustomerBase, publicVersions);
+
+    const { namespace, title, enableTypeScript } = attributeAnswers;
+    const {
+        application: { id, bspName }
+    } = configAnswers;
 
     return {
         app: {
-            id: defaults.namespace,
-            reference: configAnswers.application.id,
+            id: namespace,
+            reference: id,
             layer,
-            title: '',
+            title,
+            bspName,
             content: [getNewModelEnhanceWithChange()]
         },
         ui5: {
@@ -85,7 +91,7 @@ export async function getConfig(options: ConfigOptions): Promise<AdpWriterConfig
         target,
         options: {
             fioriTools: true,
-            enableTypeScript: false
+            enableTypeScript
         }
     };
 }