+++ packages/ui-service-inquirer/test/prompts.test.ts
@@ -91,6 +91,42 @@ const answersMock = {
     packageInputChoice: 'manualInput'
 };
 
+const getContentMock = jest.fn().mockImplementation((pckg) => {
+    const serviceName = pckg === '' ? '' : 'serviceName';
+    return Promise.resolve(
+        JSON.stringify({
+            businessService: {
+                serviceBinding: {
+                    serviceBindingName: serviceName
+                }
+            },
+            businessObject: {
+                projectionBehavior: {
+                    withDraft: undefined
+                }
+            }
+        })
+    );
+});
+
+const getContentMockDraftTrue = jest.fn().mockImplementation((pckg) => {
+    const serviceName = pckg === '' ? '' : 'serviceName';
+    return Promise.resolve(
+        JSON.stringify({
+            businessService: {
+                serviceBinding: {
+                    serviceBindingName: serviceName
+                }
+            },
+            businessObject: {
+                projectionBehavior: {
+                    withDraft: true
+                }
+            }
+        })
+    );
+});
+
 describe('getSystemQuestions', () => {
     beforeAll(() => {
         nock.disableNetConnect();
@@ -102,7 +138,7 @@ describe('getSystemQuestions', () => {
     });
 
     let questions;
-    const q: { [key: string]: Question } = {};
+    let q: { [key: string]: Question } = {};
 
     test('getSystemQuestions', async () => {
         const businessObjectMockResp = {
@@ -167,25 +203,8 @@ describe('getSystemQuestions', () => {
     });
 
     test('getServiceConfigQuestions', async () => {
-        const getContentMock = jest.fn().mockImplementation((pckg) => {
-            const serviceName = pckg === '' ? '' : 'serviceName';
-            return Promise.resolve(
-                JSON.stringify({
-                    businessService: {
-                        serviceBinding: {
-                            serviceBindingName: serviceName
-                        }
-                    },
-                    businessObject: {
-                        projectionBehavior: {
-                            withDraft: undefined
-                        }
-                    }
-                })
-            );
-        });
         const genMock = {
-            getContent: getContentMock,
+            getContent: getContentMockDraftTrue,
             validateContent: jest.fn().mockResolvedValue({
                 severity: 'OK'
             })
@@ -206,16 +225,21 @@ describe('getSystemQuestions', () => {
             objectGenerator: genMock
         } as any;
 
-        questions = getConfigPrompts(systemSelectionAnswers).prompts;
+        questions = getConfigPrompts(systemSelectionAnswers, {
+            useDraftEnabled: undefined,
+            useLaunchGen: undefined
+        }).prompts;
         questions.forEach((question) => {
             q[question.name] = question as Question;
         });
         expect(questions).toMatchSnapshot();
         expect(q.transportInputChoice.choices!()).toMatchSnapshot();
         expect(await q.serviceName.when!({ packageManual: '', packageAutocomplete: 'package' })).toEqual(true);
+        expect(PromptState.serviceConfig.showDraftEnabled).toEqual(true);
         expect(q.serviceName.choices!()).toEqual([{ name: 'serviceName', value: 'serviceName' }]);
         expect(q.serviceName.default()).toEqual(0);
         expect(await q.serviceName.validate!('testPackage')).toEqual(true);
+        expect(q.draftEnabled.when!({})).toEqual(true);
         expect(await q.draftEnabled.validate!(true)).toEqual(true);
         expect(await q.draftEnabled.validate!(false)).toEqual(true);
         expect((q.launchAppGen as YUIQuestion).additionalMessages!(false)).toBeUndefined();
@@ -223,7 +247,8 @@ describe('getSystemQuestions', () => {
             message: 'info.appGenLaunch',
             severity: 2
         });
-
+    });
+    test('getServiceConfigQuestions with error', async () => {
         const genMockValidateContent1 = {
             getContent: getContentMock,
             validateContent: jest.fn().mockResolvedValue({
@@ -247,19 +272,38 @@ describe('getSystemQuestions', () => {
         } as any;
 
         PromptState.resetConnectedSystem();
-        questions = getConfigPrompts(systemSelectionAnswers1).prompts;
+        q = {};
+        questions = getConfigPrompts(systemSelectionAnswers1, {
+            useDraftEnabled: false,
+            useLaunchGen: false
+        }).prompts;
         questions.forEach((question) => {
             q[question.name] = question as Question;
         });
         expect(await q.serviceName.when!({ packageManual: 'package' })).toEqual(true);
         expect(await q.serviceName.choices!()).toEqual([{ name: 'serviceName', value: 'serviceName' }]);
         expect(await q.serviceName.validate!('testPackage')).toMatchSnapshot();
-        expect(await q.draftEnabled.validate!(false)).toEqual('error.validatingContent');
+        expect(q.draftEnabled).toBeUndefined();
+        expect(q.launchAppGen).toBeUndefined();
         PromptState.resetServiceConfig();
         expect(await q.serviceName.when!({ packageManual: '', packageAutocomplete: '' })).toEqual(false);
         PromptState.resetServiceConfig();
         expect(await q.serviceName.when!({ packageManual: undefined, packageAutocomplete: undefined })).toEqual(false);
 
+        PromptState.resetConnectedSystem();
+        q = {};
+        questions = getConfigPrompts(systemSelectionAnswers1, {
+            useDraftEnabled: true,
+            useLaunchGen: false
+        }).prompts;
+        questions.forEach((question) => {
+            q[question.name] = question as Question;
+        });
+        expect(await q.serviceName.when!({ packageManual: 'package' })).toEqual(true);
+        expect(await q.draftEnabled.validate!(false)).toEqual('error.validatingContent');
+    });
+
+    test('getServiceConfigQuestions with error in validateContent', async () => {
         const genMockValidateContent2 = {
             getContent: getContentMock,
             validateContent: jest.fn().mockImplementation(() => {